# Update MsdFileParser