- This change makes it work a little faster

Baseline
00:46.853: ParseMsdFile took 11.715s
--------------------------------------
reducing a few calls
00:30.309: ParseMsdFile took 10.441s
--------------------------------------
removing at()
00:42.833: ParseMsdFile took 7.599s
--------------------------------------
removing gsubs
00:26.680: ParseMsdFile took 7.516s
--------------------------------------