@quietly-turning 2020-08-06
Thanks for working on this. 🙂

Can you talk about (or even just link to relevant existing prose) why the changes here improved execution time?

I see that [Lua strings are immutable](https://www.lua.org/pil/2.4.html).   Is the reasoning that:

 * reducing two consecutive`gsub` calls to one in `AddParam()` results in half the new Lua strings be allocated
 * performing `sub()` once and storing the value results in fewer allocations than constantly allocating with each call to `at()`