@teejusb 2020-08-06
I don't really have much documentation but I can walk you through my thoughts.

- Reducing a few calls
I stored some of the `sub()` results into a local variable (`char1`). I went from 14 to 11 function calls which cut down some of this time by reusing results.

- Removing `at()`
`at()` was a helper function that would just call `sub()`. Since it is a function itself, there was some function overhead which slowed down the execution. The aforementioned 14 calls was actually 28 function calls (a call to at(), which would in turn call sub()). 

- removing gsubs
We had an extra `gsub` that could be combined with previous one. This reduces one function call. I presume each call to `gsub` returns a string, so chaining the n `gsub` calls would generate n intermediate strings.