@quietly-turning 2020-08-06
Thanks.  I'm able to follow along with *what* you changed, but my understanding is more shaky when I try to understand *why* it resulted in faster execution time.

Is there a general lesson or pattern that can be distilled out of this cleanup?