@teejusb 2020-08-06
There are more instructions that need to be executed when one does many nested function calls. In compiled programs, this often gets optimized out but I think lua is not smart about it.

https://imgur.com/a/I2wOPl1

Here's an album showing the function overhead of calling functions in C++. The "call" instruction (also "jmp") is generally slower than other instructions (like add), so having a lot of nested calls can make execution slow.  The assembly itself is not the most useful part, but even just thinking of it as "number of instructions executed" is useful to gain an idea of what's going on under the hood. There's a tradeoff of readability and optimizing code here. I think in general, if you intend to use a helper function A LOT, then it might be better to get rid of the helper function itself as long as it doesn't affect readability.

In the particular case of hash generation, since we read /every single character/ in the file (many thousands/tens of thousands), taking out the at() helper function cuts down on a very large amount of extraneous function calls. For general purpose programming, where you might not be working on that scale it's generally okay.

Let me know if you have any further questions! Or if there's something I can explain better!