@quietly-turning 2020-08-07
> thinking of it as "number of instructions executed" is useful to gain an idea of what's going on under the hood

I took some months to learn MIPS in 2018 and had a flash of insight during that time into why assembly instructions matter.  It was one of the most satisfying programming experiences I've ever known as a human.

I'm curious to learn more about the relationship between Lua, the Lua interpreter, and whatever assembly results, but I'll surely die of old age before I'm ever comfortable with any of that.